refactor: test code 수정
