+++ src/test/java/com/example/demo/boundedContext/order/service/OrderDetailServiceTest.java
@@ -86,6 +86,7 @@ class OrderDetailServiceTest {
                 .build());
 
         Order order = orderRepository.save(Order.builder()
+                        .paymentKey("paymentKey")
                 .member(member)
                 .build());
